+++ src/data_access/MockAPIDAO.java
@@ -3,7 +3,6 @@ package data_access;
 import entity.CommonPlayerFactory;
 import entity.Player;
 import entity.PlayerFactory;
-import entity.Stats;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -45,11 +44,6 @@ public class MockAPIDAO implements APIinterface {
 
     @Override
     public Map<String, Object> getGeneralPlayerInfo(int id) {
-        return null;
-    }
-
-    @Override
-    public Stats getStats() {
         return null;
     }
