+++ src/entity/dummys/TeamComparatorDummyFactory.java
@@ -1,6 +1,8 @@
 package entity.dummys;
 
-public class TeamComparatorDummyFactory implements entity.TeamComparatorFactory{
+import entity.entity_helpers.TeamComparatorFactory;
+
+public class TeamComparatorDummyFactory implements TeamComparatorFactory {
     public TeamComparatorDummy create() {
         return new TeamComparatorDummy(new TeamEvaluatorDummy(new PlayerEvaluatorDummy()));
     }
