+++ src/entity/dummys/TeamEvaluatorDummy.java
@@ -1,10 +1,10 @@
 package entity.dummys;
 
-import entity.PlayerEvaluator;
+import entity.entity_helpers.PlayerEvaluator;
 import entity.Team;
-import java.util.Random;
+import entity.entity_helpers.TeamEvaluator;
 
-public class TeamEvaluatorDummy extends entity.TeamEvaluator{
+public class TeamEvaluatorDummy extends TeamEvaluator {
     public TeamEvaluatorDummy(PlayerEvaluator playerEvaluator) {
         super(playerEvaluator);
     }
@@ -13,7 +13,7 @@ public class TeamEvaluatorDummy extends entity.TeamEvaluator{
     public float evaluateTeam(Team t) {
         float teamScore = 0;
         for (entity.Player p : t.getTeamPlayers()) {
-            teamScore += playerEvaluator.evaluatePlayer(p, null);
+            teamScore += playerEvaluator.evaluatePlayer(p);
         }
         return teamScore;
     }
