+++ src/entity/entity_helpers/PlayerEvaluator.java
@@ -1,4 +1,6 @@
-package entity;
+package entity.entity_helpers;
+
+import entity.Player;
 
 /**
  * Interface (whose implementations are) used to evaluate a Player with the offered 'evaluatePlayer(Player)' method.
@@ -12,5 +14,5 @@ public interface PlayerEvaluator {
      * @param player the player to be evaluated
      * @return a float corresponding to the "evaluation" of the player
      */
-    float evaluatePlayer(Player player, Stats statsBoundary);
+    float evaluatePlayer(Player player);
 }
