+++ src/entity/entity_helpers/TeamComparatorFactory.java
@@ -1,4 +1,4 @@
-package entity;
+package entity.entity_helpers;
 
 public interface TeamComparatorFactory {
     TeamComparator create();
