+++ src/use_case/leaderboard/LeaderboardFileUserDataAccessInterface.java
@@ -1,6 +1,6 @@
 package use_case.leaderboard;
 
-import entity.TeamComparator;
+import entity.entity_helpers.TeamComparator;
 
 public interface LeaderboardFileUserDataAccessInterface {
     //The following methods are used to get the data from the database, this is a stand-in as we have not yet implemented Teams
