+++ test/use_case/leaderboard/LeaderboardTest.java
@@ -6,6 +6,9 @@ import entity.*;
 import entity.dummys.PlayerEvaluatorDummy;
 import entity.dummys.TeamComparatorDummy;
 import entity.dummys.TeamEvaluatorDummy;
+import entity.entity_helpers.PlayerEvaluator;
+import entity.entity_helpers.TeamComparator;
+import entity.entity_helpers.TeamEvaluator;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Test;
 import use_case.leaderboard.interface_adapter.LeaderboardViewModel;
